More commands, events & bug fix

+++ src/main/java/ru/kredwi/clan/commands/sub/owner/Disband.java
@@ -1,5 +1,6 @@
-package ru.kredwi.clan.commands.sub;
+package ru.kredwi.clan.commands.sub.owner;
 
+import cn.nukkit.Player;
 import cn.nukkit.command.CommandSender;
 import cn.nukkit.permission.Permission;
 import org.jspecify.annotations.NonNull;
@@ -10,16 +11,17 @@ import ru.kredwi.clan.service.ClanService;
 
 import java.util.List;
 
-public class Remove extends OwnerCommand {
-
-    public Remove(ClanService clanService) {
+public class Disband extends OwnerCommand {
+    public Disband(ClanService clanService) {
         super(clanService);
     }
 
     @Override
     protected void onCommand(@NonNull Clan clan, @NonNull CommandSender sender, @NonNull List<String> args) {
-        clanService.remove(clan.getId());
-        sender.sendMessage("Clan with id " + clan.getId() + " successfully removed");
+        if (clan.getOwnerId().equals(((Player) sender).getUniqueId())) {
+            clanService.remove(clan.getId());
+            sender.sendMessage("Clan successfully disbanded");
+        }
     }
 
     @Override
